index_recognision/pyproject.toml
[tool.poetry]
name = "IndexSystem"
version = "0.1.0"
description = ""
authors = ["Name <****@****>"]
package-mode = false 

[tool.poetry.dependencies]
python = ">=3.11,<4"
opencv-python = ">=4.12.0.88,<5.0.0.0"
python-doctr = ">=1.0.0,<2.0.0"
pika = ">=1.3.2,<2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"